pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel", "setuptools_scm[toml]>=8.1"]
build-backend = "setuptools.build_meta"

[project]
name = "homodyne"
dynamic = ["version"]
description = "JAX-first homodyne scattering analysis for XPCS under nonequilibrium conditions"
readme = "README.md"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Wei Chen", email = "weichen@anl.gov"},
    {name = "Hongrui He", email = "hhe@anl.gov"},
]
maintainers = [
    {name = "Wei Chen", email = "weichen@anl.gov"},
]
keywords = [
    "xpcs",
    "scattering",
    "dynamics",
    "jax",
    "optimization",
    "bayesian",
    "physics",
    "materials-science"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Mathematics",
]
requires-python = ">=3.10"

# Core dependencies for JAX-first architecture (2025 latest stable)
dependencies = [
    # Core scientific computing (latest stable 2025)
    "numpy>=2.0.0,<3.0.0",  # NumPy 2.3.3 latest
    "scipy>=1.14.0,<2.0.0",  # SciPy 1.16.2 latest

    # JAX ecosystem (2025 latest stable versions)
    "jax>=0.7.2,<0.8.0",  # JAX 0.7.2 latest stable
    "jaxlib>=0.7.2,<0.8.0",  # jaxlib 0.7.2 latest stable
    "optimistix>=0.0.7",  # Optimistix - modern JAX optimization library
    "equinox>=0.11.11",  # Required by optimistix

    # Bayesian inference (2025 latest)
    "numpyro>=0.18.0,<0.20.0",  # NumPyro 0.18.0 latest stable
    "blackjax>=1.2.0,<2.0.0",  # BlackJAX 1.2.5 latest

    # Data handling
    "h5py>=3.10.0,<4.0.0",  # h5py 3.14.0 latest
    "pyyaml>=6.0.2",  # PyYAML 6.0.2 latest

    # System utilities
    "psutil>=6.0.0",  # psutil 7.1.0 latest

    # Plotting (latest stable)
    "matplotlib>=3.8.0,<4.0.0",  # matplotlib 3.10.6 latest
]

[project.optional-dependencies]
# GPU acceleration using system CUDA (recommended for HPC)
gpu = [
    "jax[cuda12-local]>=0.7.2,<0.8.0",  # Uses system CUDA, avoids conflicts
    "jaxlib>=0.7.2,<0.8.0",
]

# Alternative: Bundled CUDA (may conflict with HPC environments)
gpu-bundled = [
    "jax[cuda12]>=0.7.2,<0.8.0",  # Bundles CUDA libraries
    "jaxlib>=0.7.2,<0.8.0",
]

# Future-proofing for CUDA 13
gpu-cuda13 = [
    "jax[cuda13-local]>=0.7.2,<0.8.0",  # When available
    "jaxlib>=0.7.2,<0.8.0",
]

# HPC environment (system CUDA + optimization tools)
hpc = [
    "jax[cuda12-local]>=0.7.2,<0.8.0",
    "jaxlib>=0.7.2,<0.8.0",
    "psutil>=6.0.0",
    "tqdm>=4.66.0",  # Progress bars for NLSQ
]

# Development and testing (2025 latest stable)
dev = [
    "pytest>=8.3.0,<9.0.0",  # pytest 8.3.3 latest
    "pytest-cov>=6.0.0,<8.0.0",  # pytest-cov 7.0.0 latest
    "black>=25.0.0,<26.0.0",  # black 25.9.0 latest (2025 style)
    "ruff>=0.13.0,<0.14.0",  # ruff 0.13.1 latest
    "mypy>=1.18.0,<2.0.0",  # mypy 1.18.1 latest
    "pre-commit>=4.0.0,<5.0.0",  # pre-commit 4.x latest
    "tqdm>=4.66.0",  # tqdm 4.67.x latest
]

# Documentation (2025 latest stable)
docs = [
    "sphinx>=8.0.0,<9.0.0",  # sphinx 8.3.0 latest
    "sphinx-rtd-theme>=3.0.0,<4.0.0",  # sphinx-rtd-theme 3.0.2 latest
    "myst-parser>=4.0.0,<5.0.0",  # myst-parser 4.0.1 latest
]

# Complete installation with all features
all = [
    "homodyne[hpc,dev,docs]",
    "tqdm>=4.66.0",
]

[project.urls]
Homepage = "https://github.com/your-org/homodyne"
Documentation = "https://homodyne.readthedocs.io"
Repository = "https://github.com/your-org/homodyne.git"
Issues = "https://github.com/your-org/homodyne/issues"

[project.scripts]
homodyne = "homodyne.cli.main:main"

[tool.setuptools_scm]
write_to = "homodyne/_version.py"
version_scheme = "post-release"
local_scheme = "dirty-tag"

[tool.setuptools.packages.find]
where = ["."]
include = ["homodyne*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
homodyne = [
    "config/templates/*.yaml",
    "config/templates/*.yml",
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "jax.*",
    "optimistix.*",
    "equinox.*",
    "numpyro.*",
    "blackjax.*",
    "h5py.*",
    "psutil.*",
    "matplotlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["homodyne"]
omit = [
    "homodyne/_version.py",
    "tests/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Environment-specific configurations
[tool.homodyne]
# HPC environment configuration
hpc = true
cuda_system_path = "/usr/local/cuda-12"  # CUDA 12.1-12.9 required
preferred_device = "auto"  # auto, cpu, gpu
cuda_version = "12.x"  # Minimum 12.1, maximum 12.9

# Default analysis parameters
[tool.homodyne.defaults]
analysis_mode = "static_isotropic"
optimization_method = "nlsq"
output_format = "yaml"

[tool.homodyne.performance]
# Performance optimization hints
cpu_threads = "auto"
gpu_memory_fraction = 0.9
enable_jit = true
cache_compilation = true